chore: replace `redisHashKeyValueDB` with new implementation
