+++ src/test/redis.hash.manual.test.ts
@@ -1,16 +1,18 @@
-import {
-  runCommonKeyValueDaoTest,
-  runCommonKeyValueDBTest,
-  TEST_TABLE,
-} from '@naturalcycles/db-lib/dist/testing'
+import { CommonKeyValueDao, CommonKeyValueDaoMemoCache } from '@naturalcycles/db-lib'
+import { runCommonKeyValueDBTest, TEST_TABLE } from '@naturalcycles/db-lib/dist/testing'
+import { runCommonKeyValueDaoTest } from '@naturalcycles/db-lib/dist/testing/keyValueDaoTest'
 import { KeyValueDBTuple } from '@naturalcycles/db-lib/src/kv/commonKeyValueDB'
-import { _range, localTime, pDelay } from '@naturalcycles/js-lib'
+import { _AsyncMemo, _range, localTime, pDelay } from '@naturalcycles/js-lib'
 import { RedisClient } from '../redisClient'
-import { RedisHashKeyValueDB } from '../redisHashKeyValueDB'
+import { RedishHashKeyValueDB } from '../redisHashKeyValueDB'
 
 const client = new RedisClient()
-const hashKey = 'hashField'
-const db = new RedisHashKeyValueDB({ client, hashKey })
+const db = new RedishHashKeyValueDB({ client })
+
+const dao = new CommonKeyValueDao<string, Buffer>({
+  db,
+  table: TEST_TABLE,
+})
 
 afterAll(async () => {
   await client.disconnect()
@@ -20,9 +22,10 @@ test('connect', async () => {
   await db.ping()
 })
 
-describe('runCommonHashKeyValueDBTest', () => runCommonKeyValueDBTest(db))
+describe('runCommonKeyValueDBTest', () => runCommonKeyValueDBTest(db))
 describe('runCommonKeyValueDaoTest', () => runCommonKeyValueDaoTest(db))
 
+// Saving expiring hash fields is not supported until Redis 7.4.0
 test.skip('saveBatch with EXAT', async () => {
   const testIds = _range(1, 4).map(n => `id${n}`)
   const testEntries: KeyValueDBTuple[] = testIds.map(id => [id, Buffer.from(`${id}value`)])
@@ -35,4 +38,35 @@ test.skip('saveBatch with EXAT', async () => {
   await pDelay(2000)
   loaded = await db.getByIds(TEST_TABLE, testIds)
   expect(loaded.length).toBe(0)
+})
+
+class C {
+  @_AsyncMemo({
+    cacheFactory: () =>
+      new CommonKeyValueDaoMemoCache({
+        dao,
+        ttl: 1,
+      }),
+  })
+  async get(k: string): Promise<Buffer | null> {
+    console.log(`get ${k}`)
+    return Buffer.from(k)
+  }
+}
+
+const c = new C()
+
+test('CommonKeyValueDaoMemoCache serial', async () => {
+  for (const _ of _range(10)) {
+    console.log(await c.get('key'))
+    await pDelay(100)
+  }
+})
+
+test('CommonKeyValueDaoMemoCache async swarm', async () => {
+  await Promise.all(
+    _range(30).map(async () => {
+      console.log(await c.get('key'))
+    }),
+  )
 })
